feat: add orderbook API base url (#106)
# Description
Allow to set your own Orderbook API URL instead of the default ones from
the SDK.

This PR is a follow up on
https://github.com/cowprotocol/cow-sdk/pull/177.

The motivation, is to allow to set our service IP instead of the
external one, to make the connection more performant.

## Bonus
While testing this, i realized that in case of an error, the SDK
re-attempts 10 times using exponential backoff. This makes it take too
much time in case of an error. I reduced to 5 the attempts, and now it
re-attempts, but it will take at most a few seconds instead of 2-3
minutes.

## Follow up
After this is merged, we will need to set it up (please, let me do it!) 

See

https://cowservices.slack.com/archives/C05RMJB7ZFA/p1697021044757619?thread_ts=1697007472.259279&cid=C05RMJB7ZFA

## Test
Pass the `--orderBookApi ` and make sure it uses the parameter

```bash
ts-node ./src/index.ts run --rpc <your-rpc> --dployment-block 9847564 --page-size 5000 --orderBookApi https://api.cow.fi/goerli
```

Once this is in staging we can test this in the infrastructure project.