feat: add basic auth (#121)
# Description
Implement basic auth for RPCs

This PR will make worse the problem we discussed about the chain config
being too error prone. So Im glad of this, so we make it more obvious
this needs to change (into individual params or a config file) --> out
the scope

The PR adds also support for HTTP (but this should not be used in the
wild internet!! just used to test and for some internal RPC that don't
have HTTPS. Note that basic auth and HTTPS is a no-no!

# CTX
The motivation of this PR is that GC RPC endpoint started to fail. See
https://cowservices.slack.com/archives/C0361CDD1FZ/p1700845414302179


# Test 
Just add user/password at the end. Credentials shared in 1password

Example:
```bash
yarn ts-node ./src/index.ts run --chain-config <rpc>,,,,https://raw.githubusercontent.com/cowprotocol/watch-tower/config/staging/filter-policy-100.json,<user>,<password> \ 
  --page-size 5000
```