Fix Example Configuration File
When running with the example configuration file, you get the following
error:

```
Error parsing configuration file: SyntaxError: Expected double-quoted property name in JSON at position 387 (line 11 column 17)
    at JSON.parse (<anonymous>)
    at validateJSON (/usr/src/app/dist/src/index.js:122:27)
    at Command.<anonymous> (/usr/src/app/dist/src/index.js:65:24)
    at Command.listener [as _actionHandler] (/usr/src/app/node_modules/commander/lib/command.js:482:17)
    at /usr/src/app/node_modules/commander/lib/command.js:1300:65
    at Command._chainOrCall (/usr/src/app/node_modules/commander/lib/command.js:1197:12)
    at Command._parseCommand (/usr/src/app/node_modules/commander/lib/command.js:1300:27)
    at /usr/src/app/node_modules/commander/lib/command.js:1081:27
    at Command._chainOrCall (/usr/src/app/node_modules/commander/lib/command.js:1197:12)
    at Command._dispatchSubcommand (/usr/src/app/node_modules/commander/lib/command.js:1077:23)
```

It appears that it doesn't like the trailing `,`s (so it appears to only
support plain old JSON and not JSONc or JSON5).

This PR fixes the example configuration file to be valid JSON (and also
removes some trailing whitespace in the file).