feat: show id of order and tx hash when posting the discrete order (#169)
# Description
This PR is the most pressing change in the proposal to improve
debuggability of watch-tower described in
https://docs.google.com/document/d/1rNM4pqS-d17Yx1Ei_RMMvER5vJfK1c8I7Ik74PbOHSc/edit?tab=t.0


Basically, this PR allows us to connect the discrete orders with the
programmatic order that created them. This is really important, because
we can connect some issues on some orders that are wrongly created by a
smart contract, with the smart contract that created them in the first
place.