feat: apply filter policy before adding an order (#176)
# Description
We are not respecting the filter policy in order to add or not add
orders from events.


## Context
I noticed how staging indexed this an order of an owner that shouldn't
have. However, in the next block got dropped because we apply the
filtering using the policies at the beginning of processing a block.

This explain this increment + decreement:

![image](https://github.com/user-attachments/assets/f92eb74c-1ab0-4d15-8002-e3cbd8f7b08b)


You can also see this in the logs:

![image](https://github.com/user-attachments/assets/ed7c9c34-f4b9-4e94-8602-f591b503974f)


## Chages
This pr applies the filter policy before adding an order, so we don't
add it in the first place.

## Test
You can create one TWAP using any account and see how now we don't get
this increment and decrement in Grafana.